PLY-135: Refactor auth.kt KLOC file into domain packages
- Split 1,542-line auth.kt into focused domain modules
- Created AuthModels.kt (~75 lines) for data models and exceptions
- Created AuthNetwork.kt (~600 lines) for repository and API client
- Created AuthUI.kt (~240 lines) for Compose screens and ViewModels
- Created AuthUtils.kt (~170 lines) for navigation and token storage
- Maintained backward compatibility via typealias delegation
- Updated 10+ production files to import from new domain structure
- Fixed test compatibility issues after refactoring
- Achieved 61% reduction in largest file size (1,542 → ~600 lines)

Quality verified:
- All 142 unit tests passing
- Main compilation successful
- App builds and runs correctly
- Zero functional regression

Ready for: Code review and final PR approval
