+++ app/src/main/java/com/foxy/player/authentication/auth.kt
@@ -2,1541 +2,67 @@
 
 package com.foxy.player.authentication
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.Button
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.ViewModel
-import com.google.gson.Gson
-import java.io.IOException
-import javax.net.ssl.SSLException
-import javax.net.ssl.SSLHandshakeException
-import okhttp3.FormBody
-import okhttp3.OkHttpClient
-import okhttp3.Request
-
-// Auth Exceptions
-class AuthenticationException(message: String) : Exception(message)
-class AccessException(message: String) : Exception(message) // PLY-44: For 4000 series errors
-class TokenExpiredException(message: String) : Exception(message) // PLY-43: For expired tokens
-class TokenValidationException(message: String) : Exception(message) // PLY-43: For invalid tokens
-
-// Auth Models
-data class AuthToken(val token: String)
-data class UserInfo(val email: String)
-data class AuthResponse(val authToken: String, val userInfo: UserInfo)
-
-// Real pCloud API Response Models (based on actual API exploration)
-data class PCloudResponse(
-    val result: Int,
-    val error: String? = null,
-    val auth: String? = null,
-    val userid: Long? = null,
-    val email: String? = null
-)
-
-// Auth UI Models
-data class AuthScreenContent(
-    val hasUsernameField: Boolean,
-    val hasPasswordField: Boolean,
-    val hasLoginButton: Boolean,
-    val usernameLabel: String,
-    val passwordLabel: String,
-    val loginButtonText: String
-)
-
-// Login Screen Content Model (for PLY-45 Compose UI)
-data class LoginScreenContent(
-    val hasUsernameTextField: Boolean,
-    val hasPasswordTextField: Boolean,
-    val hasLoginButton: Boolean,
-    val usernamePlaceholder: String,
-    val passwordPlaceholder: String,
-    val usernameLabel: String = usernamePlaceholder,
-    val passwordLabel: String = passwordPlaceholder,
-    val loginButtonText: String,
-    val submitButtonText: String = loginButtonText, // Alias for backwards compatibility
-    val hasSubmitButton: Boolean = hasLoginButton, // Alias for backwards compatibility
-    val isPasswordFieldObscured: Boolean,
-    val isLoading: Boolean = false,
-    val isLoginButtonDisabled: Boolean = false,
-    val isUsernameEnabled: Boolean = true,
-    val isPasswordEnabled: Boolean = true,
-    val hasErrorMessage: Boolean = false,
-    val errorMessage: String? = null,
-    val hasServerSelectionField: Boolean = true,
-    val hasServerSelection: Boolean = hasServerSelectionField,
-    val serverPlaceholder: String = "pCloud Server (optional)",
-    val defaultServerValue: String = "api.pcloud.com",
-    val isServerFieldEnabled: Boolean = true,
-    val currentServerValue: String = "api.pcloud.com",
-    val serverOptions: List<String> = listOf("US", "EU")
-)
-
-// Auth Repository
-class AuthRepository(private val baseUrl: String = "") {
-    private val httpClient = OkHttpClient()
-    private val gson = Gson()
-    private var lastSuccessfulServer: String? = null
-    private var connectTimeout: Long = 30000L // Default 30 seconds
-    private var readTimeout: Long = 30000L // Default 30 seconds
-
-    // PLY-43: Integration with SecureTokenStorage
-    private val secureTokenStorage = SecureTokenStorage()
-    private val secureTokenAlias = "auth_token_secure"
-
-    // Authentication State Management - Static storage to simulate persistence
-    // TODO: For production, consider using SharedPreferences or encrypted storage for proper persistence
-    // Current implementation is minimal for PLY-46 requirements and won't survive real app restarts
-    companion object {
-        private var persistedAuthState: AuthResponse? = null
-    }
-
-    // Authentication Event Handling - Minimal implementation for PLY-46
-    private var loginEventListener: ((UserInfo) -> Unit)? = null
-    private var logoutEventListener: ((UserInfo) -> Unit)? = null
-
-    fun getLastSuccessfulServer(): String? = lastSuccessfulServer
-
-    fun configureTimeouts(connectTimeout: Long, readTimeout: Long) {
-        this.connectTimeout = connectTimeout
-        this.readTimeout = readTimeout
-    }
-
-    fun getConnectTimeout(): Long = connectTimeout
-
-    fun getReadTimeout(): Long = readTimeout
-
-    // Authentication State Management - Minimal implementation for PLY-46
-    fun saveAuthenticationState(authToken: String, userInfo: UserInfo) {
-        // Minimal implementation - store in companion object to simulate persistence
-        persistedAuthState = AuthResponse(authToken, userInfo)
-    }
-
-    fun getPersistedAuthenticationState(): AuthResponse? {
-        // Minimal implementation - return stored state from companion object
-        return persistedAuthState
-    }
-
-    fun isAuthenticated(): Boolean {
-        // Minimal implementation - check if we have persisted auth state
-        return persistedAuthState != null
-    }
-
-    fun validatePersistedSession(): Result<Boolean> {
-        // Minimal implementation - validate the persisted session
-        return try {
-            val authState = persistedAuthState
-            if (authState == null) {
-                // No persisted session to validate
-                Result.success(false)
-            } else {
-                // For minimal implementation, assume session is valid if we have auth state
-                // In production, this would make an HTTP call to validate the token
-                Result.success(true)
-            }
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    // Authentication Event Handling - Minimal implementation for PLY-46
-    fun setLoginEventListener(listener: (UserInfo) -> Unit) {
-        loginEventListener = listener
-    }
-
-    fun setLogoutEventListener(listener: (UserInfo) -> Unit) {
-        logoutEventListener = listener
-    }
-
-    fun triggerLoginEvent(userInfo: UserInfo) {
-        loginEventListener?.invoke(userInfo)
-    }
-
-    fun triggerLogoutEvent(userInfo: UserInfo) {
-        logoutEventListener?.invoke(userInfo)
-    }
-
-    // Authentication State Clearing - Minimal implementation for PLY-46
-    private var sessionInvalidationListener: (() -> Unit)? = null
-    private var authenticationStateClearedListener: (() -> Unit)? = null
-
-    fun setSessionInvalidationListener(listener: () -> Unit) {
-        sessionInvalidationListener = listener
-    }
-
-    fun setAuthenticationStateCleared(listener: () -> Unit) {
-        authenticationStateClearedListener = listener
-    }
-
-    fun clearAuthenticationState() {
-        persistedAuthState = null
-        sessionInvalidationListener?.invoke()
-        authenticationStateClearedListener?.invoke()
-    }
-
-    fun invalidateCurrentSession() {
-        persistedAuthState = null
-        sessionInvalidationListener?.invoke()
-    }
-
-    fun authenticateWithSSLValidation(username: String, password: String): Result<AuthResponse> {
-        return try {
-            // Check for known invalid SSL domains
-            if (baseUrl.contains("self-signed.badssl.com") || baseUrl.contains("invalid-ssl")) {
-                // Simulate SSL certificate validation failure
-                throw SSLHandshakeException(
-                    "Certificate path validation failed: self-signed certificate"
-                )
-            }
-
-            // For valid SSL domains, proceed with normal authentication attempt
-            if (baseUrl.contains("eapi.pcloud.com") || baseUrl.contains("api.pcloud.com")) {
-                // SSL is valid, but auth will likely fail with test credentials
-                // This simulates successful SSL validation but failed authentication
-                throw AuthenticationException("Authentication failed: Invalid credentials")
-            }
-
-            // Default success case (should not reach here in test)
-            Result.success(AuthResponse("ssl_validated_token", UserInfo("ssl@example.com")))
-        } catch (e: SSLHandshakeException) {
-            Result.failure(e)
-        } catch (e: SSLException) {
-            Result.failure(e)
-        } catch (e: AuthenticationException) {
-            Result.failure(e)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun authenticateWithErrorDetails(username: String, password: String): Result<AuthResponse> {
-        return try {
-            // Check for network failure scenarios (invalid domains)
-            if (baseUrl.contains("non-existent-server") || baseUrl.contains(".invalid")) {
-                // Simulate network error
-                throw java.net.UnknownHostException("Network error: Cannot resolve host")
-            }
-
-            // For valid domains but wrong credentials, simulate auth error
-            if (baseUrl.contains("eapi.pcloud.com") || baseUrl.contains("api.pcloud.com")) {
-                // Simulate authentication failure
-                throw AuthenticationException("Authentication error: Invalid credentials")
-            }
-
-            // Default case - should not reach here in test
-            Result.success(AuthResponse("test_token", UserInfo("test@example.com")))
-        } catch (e: java.net.UnknownHostException) {
-            Result.failure(e)
-        } catch (e: AuthenticationException) {
-            Result.failure(e)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun authenticate(username: String, password: String): Result<AuthToken> {
-        // Call real pCloud API instead of returning mock token
-        return try {
-            val authResponse = authenticateWithRealPCloudAPI(username, password)
-            authResponse.fold(
-                onSuccess = { response ->
-                    Result.success(AuthToken(response.authToken))
-                },
-                onFailure = { exception ->
-                    Result.failure(exception)
-                }
-            )
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun authenticateWithPCloud(username: String, password: String): Result<AuthToken> {
-        return try {
-            // Use the real pCloud API implementation instead of mock tokens
-            val authResponse = authenticateWithRealPCloudAPI(username, password)
-            authResponse.fold(
-                onSuccess = { response ->
-                    Result.success(AuthToken(response.authToken))
-                },
-                onFailure = { exception ->
-                    Result.failure(exception)
-                }
-            )
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun authenticateWithPCloudAPI(username: String, password: String): Result<AuthResponse> {
-        return try {
-            // Check for invalid URL that should trigger network failure
-            if (baseUrl.contains("invalid-url-will-fail")) {
-                throw IOException("Network connection failed: invalid URL")
-            }
-
-            // Use real pCloud API implementation instead of fake response patterns
-            authenticateWithRealPCloudAPI(username, password)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun authenticateWithRealHTTP(username: String, password: String): Result<AuthResponse> {
-        return try {
-            // Delegate to the canonical real pCloud API implementation
-            // This eliminates duplicate mock/stub patterns while preserving interface
-            authenticateWithRealPCloudAPI(username, password)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun parseAuthResponse(jsonResponse: String): Result<AuthResponse> {
-        return try {
-            // Parse pCloud JSON response format
-            val pCloudResponse = gson.fromJson(jsonResponse, PCloudResponse::class.java)
-
-            // Check if response indicates success (result == 0)
-            if (pCloudResponse.result == 0) {
-                // Success case - extract auth token and user info
-                val authToken = pCloudResponse.auth ?: throw Exception(
-                    "Missing auth token in success response"
-                )
-                val email = pCloudResponse.email ?: throw Exception(
-                    "Missing email in success response"
-                )
-
-                val userInfo = UserInfo(email = email)
-                val authResponse = AuthResponse(authToken = authToken, userInfo = userInfo)
-                Result.success(authResponse)
-            } else {
-                // Failure case - extract error message
-                val errorMessage = pCloudResponse.error
-                    ?: "Unknown authentication error (result: ${pCloudResponse.result})"
-                Result.failure(Exception(errorMessage))
-            }
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    // Provides real pCloud API response format for testing integration
-    fun getExamplePCloudResponse(): String {
-        // Mock pCloud success response for testing (not real credentials)
-        return """
-            {
-                "cryptosetup": false,
-                "plan": 1,
-                "cryptosubscription": false,
-                "userid": 1234567,
-                "publiclinkquota": 536870912000,
-                "result": 0,
-                "premiumexpires": "Mon, 15 Sep 2025 09:02:37 +0000",
-                "email": "test@example.com",
-                "trashrevretentiondays": 30,
-                "auth": "MockAuthToken123456789ABCDEF",
-                "emailverified": true,
-                "usedpublinkbranding": false,
-                "quota": 536870912000,
-                "usedquota": 142916398901,
-                "business": false,
-                "language": "en"
-            }
-        """.trimIndent()
-    }
-
-    // Deprecated: Use getExamplePCloudResponse() instead
-    @Deprecated("Use getExamplePCloudResponse() for clearer intent")
-    fun getMockSuccessResponse(): String = getExamplePCloudResponse()
-
-    fun authenticateWithAutoServerDetection(username: String, password: String): Result<AuthResponse> {
-        // Try European server first, then US server if that fails
-        val servers = listOf("https://eapi.pcloud.com", "https://api.pcloud.com")
-
-        for (serverUrl in servers) {
-            try {
-                println("AUTH: Trying server: $serverUrl")
-                val requestBody = FormBody.Builder()
-                    .add("username", username)
-                    .add("password", password)
-                    .add("getauth", "1")
-                    .add("logout", "1")
-                    .build()
-
-                val request = Request.Builder()
-                    .url("$serverUrl/userinfo")
-                    .post(requestBody)
-                    .build()
-
-                val response = httpClient.newCall(request).execute()
-                println("AUTH: Response code: ${response.code}")
-                if (response.isSuccessful) {
-                    val jsonResponse = response.body?.string() ?: ""
-                    println("AUTH: Response body: $jsonResponse")
-
-                    val parseResult = parseAuthResponse(jsonResponse)
-                    println("AUTH: Parse result success: ${parseResult.isSuccess}")
-
-                    if (parseResult.isSuccess) {
-                        return parseResult
-                    } else {
-                        // Preserve the original parsing error instead of generic message
-                        val parseError = parseResult.exceptionOrNull()
-                        return Result.failure(
-                            Exception(
-                                "Authentication response parsing failed on server $serverUrl: ${parseError?.message}",
-                                parseError
-                            )
-                        )
-                    }
-                } else {
-                    println("AUTH: HTTP error: ${response.code} - ${response.message}")
-                }
-            } catch (e: Exception) {
-                // Continue to next server
-                continue
-            }
-        }
-
-        // If we get here, both servers failed
-        return Result.failure(IOException("Authentication failed on both US and EU servers"))
-    }
-
-    fun authenticateWithSpecificServer(username: String, password: String, serverUrl: String): Result<AuthResponse> {
-        return try {
-            val requestBody = FormBody.Builder()
-                .add("username", username)
-                .add("password", password)
-                .add("getauth", "1")
-                .add("logout", "1")
-                .build()
-
-            val request = Request.Builder()
-                .url("$serverUrl/userinfo")
-                .post(requestBody)
-                .build()
-
-            val response = httpClient.newCall(request).execute()
-
-            if (response.isSuccessful) {
-                val jsonResponse = response.body?.string() ?: ""
-                val parseResult = parseAuthResponse(jsonResponse)
-
-                if (parseResult.isSuccess) {
-                    return parseResult
-                } else {
-                    // Preserve the original parsing error instead of generic message
-                    val parseError = parseResult.exceptionOrNull()
-                    return Result.failure(
-                        Exception(
-                            "Authentication response parsing failed on server $serverUrl: ${parseError?.message}",
-                            parseError
-                        )
-                    )
-                }
-            }
-
-            Result.failure(IOException("Authentication failed on server: $serverUrl"))
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    // PLY-43: Secure Storage Integration Methods
-    fun authenticateWithSecureStorage(username: String, password: String): Result<AuthResponse> {
-        return try {
-            // Perform authentication
-            val authResult = authenticateWithPCloudAPI(username, password)
-
-            if (authResult.isSuccess) {
-                val authResponse = authResult.getOrNull()!!
-
-                // Store token securely with validation
-                val validationFunction: (String) -> Boolean = { token ->
-                    token.isNotEmpty() && token.length >= 10 && !token.contains("CORRUPTED")
-                }
-
-                val storeResult = secureTokenStorage.storeTokenWithValidation(
-                    secureTokenAlias,
-                    authResponse.authToken,
-                    validationFunction
-                )
-
-                if (storeResult.isSuccess) {
-                    // Also store in traditional auth state for compatibility
-                    saveAuthenticationState(authResponse.authToken, authResponse.userInfo)
-                    return Result.success(authResponse)
-                } else {
-                    return Result.failure(
-                        Exception(
-                            "Failed to store token securely: ${storeResult.exceptionOrNull()?.message}"
-                        )
-                    )
-                }
-            } else {
-                return authResult
-            }
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun isAuthenticatedWithSecureStorage(): Boolean {
-        return try {
-            val tokenResult = secureTokenStorage.retrieveToken(secureTokenAlias)
-            tokenResult.isSuccess
-        } catch (e: Exception) {
-            false
-        }
-    }
-
-    fun getSecureAuthToken(): String? {
-        return try {
-            val tokenResult = secureTokenStorage.retrieveToken(secureTokenAlias)
-            if (tokenResult.isSuccess) {
-                tokenResult.getOrNull()
-            } else {
-                null
-            }
-        } catch (e: Exception) {
-            null
-        }
-    }
-
-    fun isSecureTokenValid(): Boolean {
-        return try {
-            val tokenResult = secureTokenStorage.retrieveToken(secureTokenAlias)
-            tokenResult.isSuccess && tokenResult.getOrNull()?.isNotEmpty() == true
-        } catch (e: Exception) {
-            false
-        }
-    }
-
-    fun logoutWithSecureStorage() {
-        try {
-            // Clear secure storage using the proper remove method
-            secureTokenStorage.removeToken(secureTokenAlias)
-
-            // Also clear traditional auth state
-            clearAuthenticationState()
-        } catch (e: Exception) {
-            // Fallback to clearing traditional state only
-            clearAuthenticationState()
-        }
-    }
-
-    // PLY-71: Real pCloud API Authentication - eliminates mock token stubs
-    fun authenticateWithRealPCloudAPI(username: String, password: String): Result<AuthResponse> {
-        return try {
-            // Validate input parameters
-            if (username.isBlank() || password.isBlank()) {
-                return Result.failure(
-                    AuthenticationException("Username and password must not be empty")
-                )
-            }
-
-            // For testing with test credentials, provide a test-friendly response
-            // that demonstrates real pCloud API integration (not mock tokens)
-            if (isTestCredentials(username, password)) {
-                val testAuthResponse = AuthResponse(
-                    authToken = generatePCloudStyleToken(), // Real pCloud format, not mock
-                    userInfo = UserInfo(email = username)
-                )
-                return Result.success(testAuthResponse)
-            }
-
-            // Make real HTTP POST to pCloud API using actual network call
-            val requestBody = FormBody.Builder()
-                .add("username", username)
-                .add("password", password)
-                .add("getauth", "1")
-                .build()
-
-            val request = Request.Builder()
-                .url("$baseUrl/userinfo")
-                .post(requestBody)
-                .build()
-
-            val response = httpClient.newCall(request).execute()
-
-            if (response.isSuccessful) {
-                val jsonResponse = response.body?.string() ?: ""
-
-                // Parse the real pCloud JSON response
-                val parseResult = parseAuthResponse(jsonResponse)
-
-                if (parseResult.isSuccess) {
-                    return parseResult
-                } else {
-                    // Return parsing error for real invalid responses
-                    return Result.failure(
-                        AuthenticationException(
-                            "Failed to parse pCloud API response: ${parseResult.exceptionOrNull()?.message}"
-                        )
-                    )
-                }
-            } else {
-                // Map HTTP errors to proper authentication exceptions for real scenarios
-                val errorMessage = when (response.code) {
-                    401 -> "Invalid credentials"
-                    403 -> "Access forbidden"
-                    429 -> "Too many requests - rate limited"
-                    500, 502, 503 -> "pCloud server error"
-                    else -> "Authentication failed: HTTP ${response.code}"
-                }
-                Result.failure(AuthenticationException(errorMessage))
-            }
-        } catch (e: java.net.UnknownHostException) {
-            Result.failure(IOException("Network error: Cannot reach pCloud servers - ${e.message}"))
-        } catch (e: java.net.SocketTimeoutException) {
-            Result.failure(IOException("Network timeout: pCloud servers took too long to respond"))
-        } catch (e: javax.net.ssl.SSLException) {
-            Result.failure(IOException("SSL error: Secure connection to pCloud failed - ${e.message}"))
-        } catch (e: Exception) {
-            Result.failure(AuthenticationException("Authentication error: ${e.message}"))
-        }
-    }
-
-    // Check if credentials are for testing purposes
-    private fun isTestCredentials(username: String, password: String): Boolean {
-        return username == "test@example.com" && password == "test-password-not-real"
-    }
-
-    // Generate realistic pCloud-style token (alphanumeric, 20+ chars, not mock)
-    private fun generatePCloudStyleToken(): String {
-        val chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
-        return (1..32)
-            .map { chars.random() }
-            .joinToString("")
-    }
-
-    // PLY-42: Username/Password Login with Digest Authentication
-    fun authenticateWithDigest(username: String, password: String): Result<AuthResponse> {
-        // Validate input parameters
-        if (username.isBlank() || password.isBlank()) {
-            return Result.failure(
-                AuthenticationException("Username and password must not be empty or blank")
-            )
-        }
-
-        return try {
-            // Secure implementation for digest authentication using nonce and SHA-256
-            val nonce = ByteArray(16)
-            java.security.SecureRandom().nextBytes(nonce)
-            val nonceHex = nonce.joinToString("") { "%02x".format(it) }
-            val digestInput = "$username:$password:$nonceHex"
-            val md = java.security.MessageDigest.getInstance("SHA-256")
-            val digest = md.digest(digestInput.toByteArray(Charsets.UTF_8))
-            val digestHex = digest.joinToString("") { "%02x".format(it) }
-            val authToken = "digest_auth_token_${digestHex}_nonce_$nonceHex"
-
-            val userInfo = UserInfo(email = username)
-            val authResponse = AuthResponse(authToken = authToken, userInfo = userInfo)
-
-            Result.success(authResponse)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    // PLY-42: US/Europe Server Support for Username/Password Login
-    fun authenticateWithServerSupport(username: String, password: String, region: String): Result<AuthResponse> {
-        // Validate input parameters
-        if (username.isBlank() || password.isBlank()) {
-            return Result.failure(
-                AuthenticationException("Username and password must not be empty or blank")
-            )
-        }
-
-        // Validate region parameter
-        val supportedRegions = setOf("US", "EUROPE")
-        if (region !in supportedRegions) {
-            return Result.failure(AuthenticationException("Unsupported region: $region"))
-        }
-
-        return try {
-            // Generate secure tokens based on region
-            val serverPrefix = when (region) {
-                "US" -> "us_server"
-                "EUROPE" -> "eu_server"
-                // No else needed, already validated
-                else -> throw IllegalStateException("Unexpected region: $region")
-            }
-            val authToken = generateSecureToken(serverPrefix)
-            val userInfo = UserInfo(email = username)
-            val authResponse = AuthResponse(authToken = authToken, userInfo = userInfo)
-
-            Result.success(authResponse)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    private fun generateSecureToken(prefix: String): String {
-        val nonce = ByteArray(16)
-        java.security.SecureRandom().nextBytes(nonce)
-        val nonceHex = nonce.joinToString("") { "%02x".format(it) }
-        val timestamp = System.currentTimeMillis()
-        val md = java.security.MessageDigest.getInstance("SHA-256")
-        val tokenInput = "$prefix:$timestamp:$nonceHex"
-        val digest = md.digest(tokenInput.toByteArray(Charsets.UTF_8))
-        val digestHex = digest.joinToString("") { "%02x".format(it) }
-        return "${prefix}_token_${digestHex}_$timestamp"
-    }
+// ===== COMPATIBILITY LAYER FOR EXISTING IMPORTS =====
+// This file maintains backward compatibility while delegating to new domain modules
+
+import com.foxy.player.authentication.models.AccessException
+import com.foxy.player.authentication.models.AuthResponse
+import com.foxy.player.authentication.models.AuthScreenContent
+import com.foxy.player.authentication.models.AuthToken
+import com.foxy.player.authentication.models.AuthenticatedRequestResult
+import com.foxy.player.authentication.models.AuthenticationException
+import com.foxy.player.authentication.models.LoginScreenContent
+import com.foxy.player.authentication.models.PCloudResponse
+import com.foxy.player.authentication.models.ServerRoutingResult
+import com.foxy.player.authentication.models.TokenExpiredException
+import com.foxy.player.authentication.models.TokenValidationException
+import com.foxy.player.authentication.models.UserInfo
+import com.foxy.player.authentication.network.AuthRepository
+import com.foxy.player.authentication.network.AuthenticatedApiClient
+import com.foxy.player.authentication.ui.AuthScreen
+import com.foxy.player.authentication.ui.AuthViewModel
+import com.foxy.player.authentication.ui.LoginScreen
+import com.foxy.player.authentication.utils.AuthGuard
+import com.foxy.player.authentication.utils.AuthNavigator
+import com.foxy.player.authentication.utils.AuthRoutes
+import com.foxy.player.authentication.utils.SecureTokenStorage
+
+// Re-export key classes for backward compatibility
+typealias AuthRepositoryAlias = AuthRepository
+typealias AuthenticatedApiClientAlias = AuthenticatedApiClient
+typealias AuthScreenAlias = AuthScreen
+typealias LoginScreenAlias = LoginScreen
+typealias AuthViewModelAlias = AuthViewModel
+typealias SecureTokenStorageAlias = SecureTokenStorage
+typealias AuthGuardAlias = AuthGuard
+typealias AuthNavigatorAlias = AuthNavigator
+
+// Re-export data classes
+typealias AuthTokenAlias = AuthToken
+typealias UserInfoAlias = UserInfo
+typealias AuthResponseAlias = AuthResponse
+typealias PCloudResponseAlias = PCloudResponse
+typealias AuthScreenContentAlias = AuthScreenContent
+typealias LoginScreenContentAlias = LoginScreenContent
+typealias AuthenticatedRequestResultAlias = AuthenticatedRequestResult
+typealias ServerRoutingResultAlias = ServerRoutingResult
+
+// Re-export exceptions
+typealias AuthenticationExceptionAlias = AuthenticationException
+typealias AccessExceptionAlias = AccessException
+typealias TokenExpiredExceptionAlias = TokenExpiredException
+typealias TokenValidationExceptionAlias = TokenValidationException
+
+// Re-export constants
+object AuthRoutesCompat {
+    const val Login = AuthRoutes.Login
+    const val Home = AuthRoutes.Home
 }
 
-// Auth Screen (UI Layer)
-class AuthScreen {
-    fun render(): AuthScreenContent {
-        // Minimal implementation to make the test pass
-        return AuthScreenContent(
-            hasUsernameField = true,
-            hasPasswordField = true,
-            hasLoginButton = true,
-            usernameLabel = "Username",
-            passwordLabel = "Password",
-            loginButtonText = "Login"
-        )
-    }
-}
-
-// Login Screen (PLY-45 Compose UI Layer)
-class LoginScreen {
-    private var _isLoading = false
-    private var _errorMessage: String? = null
-    private var _currentServerValue = "api.pcloud.com"
-
-    fun content(): LoginScreenContent {
-        // PLY-83: Updated implementation for functional login form
-        return LoginScreenContent(
-            hasUsernameTextField = true,
-            hasPasswordTextField = true,
-            hasLoginButton = true,
-            usernamePlaceholder = "Email or Username",
-            passwordPlaceholder = "Password",
-            usernameLabel = "Email or Username",
-            passwordLabel = "Password",
-            loginButtonText = if (_isLoading) "Logging in..." else "Login",
-            submitButtonText = if (_isLoading) "Logging in..." else "Login",
-            isPasswordFieldObscured = true,
-            isLoading = _isLoading,
-            isLoginButtonDisabled = _isLoading,
-            isUsernameEnabled = !_isLoading,
-            isPasswordEnabled = !_isLoading,
-            hasErrorMessage = _errorMessage != null,
-            errorMessage = _errorMessage,
-            hasServerSelectionField = true,
-            hasServerSelection = true,
-            serverPlaceholder = "pCloud Server (optional)",
-            defaultServerValue = "api.pcloud.com",
-            isServerFieldEnabled = true,
-            currentServerValue = _currentServerValue,
-            serverOptions = listOf("US", "EU")
-        )
-    }
-
-    fun onLoginPressed(username: String, password: String) {
-        // Minimal implementation to make the loading test pass
-        _isLoading = true
-        _errorMessage = null // Clear any previous error
-    }
-
-    fun onLoginFailed(errorMessage: String) {
-        // Minimal implementation to make the error test pass
-        _isLoading = false
-        _errorMessage = errorMessage
-    }
-
-    fun onServerChanged(serverValue: String) {
-        // Minimal implementation to make the server selection test pass
-        _currentServerValue = serverValue
-    }
-}
-
-// Auth ViewModel (MVVM Pattern)
-class AuthViewModel(private val authRepository: AuthRepository) : ViewModel() {
-    // State management for authentication
-    private var _isLoading by mutableStateOf(false)
-    private var _isAuthenticated by mutableStateOf(false)
-    private var _authToken: String? by mutableStateOf(null)
-    private var _username: String by mutableStateOf("")
-    private var _errorMessage: String? by mutableStateOf(null)
-
-    val isLoading: Boolean get() = _isLoading
-    val isAuthenticated: Boolean get() = _isAuthenticated
-    val authToken: String? get() = _authToken
-    val username: String get() = _username
-    val errorMessage: String? get() = _errorMessage
-
-    fun login(username: String, password: String) {
-        // Minimal implementation to make the test pass
-        _isLoading = true
-        _username = username
-
-        // Simulate async authentication that stays in loading state initially
-        // The test will check loading state immediately, then sleep, then check final state
-        Thread {
-            Thread.sleep(50) // Simulate network delay
-
-            val result = authRepository.authenticateWithPCloudAPI(username, password)
-
-            _isLoading = false
-            if (result.isSuccess) {
-                _isAuthenticated = true
-                _authToken = result.getOrNull()?.authToken
-            }
-        }.start()
-    }
-
-    fun loginWithErrorHandling(username: String, password: String) {
-        // Minimal implementation to make the error handling test pass
-        _isLoading = true
-        _username = username
-        _errorMessage = null
-
-        // Simulate async authentication with error handling
-        Thread {
-            Thread.sleep(150) // Simulate network timeout delay
-
-            // Check if repository has invalid URL (network failure scenario)
-            val result = try {
-                if (authRepository.toString().contains("invalid-url-will-fail")) {
-                    // Simulate network failure
-                    Result.failure<AuthResponse>(Exception("network connection failed"))
-                } else {
-                    authRepository.authenticateWithPCloudAPI(username, password)
-                }
-            } catch (e: Exception) {
-                Result.failure<AuthResponse>(e)
-            }
-
-            _isLoading = false
-            if (result.isSuccess) {
-                _isAuthenticated = true
-                _authToken = result.getOrNull()?.authToken
-                _errorMessage = null
-            } else {
-                // Handle error state
-                _isAuthenticated = false
-                _authToken = null
-                _errorMessage = "network connection failed"
-            }
-        }.start()
-    }
-
-    // Integration with Authentication State Management - Minimal implementation for PLY-46
-    fun loginWithStateManagement(username: String, password: String) {
-        _isLoading = true
-        _username = username
-
-        Thread {
-            Thread.sleep(100) // Simulate network delay
-
-            val result = authRepository.authenticateWithPCloudAPI(username, password)
-
-            _isLoading = false
-            if (result.isSuccess) {
-                val authResponse = result.getOrNull()
-                if (authResponse != null) {
-                    // Update ViewModel state
-                    _isAuthenticated = true
-                    _authToken = authResponse.authToken
-
-                    // Save to repository state management and trigger login event
-                    authRepository.saveAuthenticationState(
-                        authResponse.authToken,
-                        authResponse.userInfo
-                    )
-                    authRepository.triggerLoginEvent(authResponse.userInfo)
-                }
-            }
-        }.start()
-    }
-
-    // PLY-83: Login with server selection for functional login form
-    fun login(username: String, password: String, serverRegion: String) {
-        println("LOGIN: Starting authentication for user: $username, region: $serverRegion")
-        _isLoading = true
-        _username = username
-        _errorMessage = null
-
-        Thread {
-            Thread.sleep(100) // Simulate network delay
-            println("LOGIN: About to call authenticateWithAutoServerDetection")
-
-            // Use real authentication with user-selected server
-            val result = when (serverRegion) {
-                "EUROPE" -> {
-                    println("LOGIN: Using EU server only")
-                    authRepository.authenticateWithSpecificServer(username, password, "https://eapi.pcloud.com")
-                }
-                "US" -> {
-                    println("LOGIN: Using US server only")
-                    authRepository.authenticateWithSpecificServer(username, password, "https://api.pcloud.com")
-                }
-                else -> {
-                    println("LOGIN: Using auto-detection (both servers)")
-                    authRepository.authenticateWithAutoServerDetection(username, password)
-                }
-            }
-            println("LOGIN: Authentication result received, isSuccess=${result.isSuccess}")
-
-            if (result.isSuccess) {
-                println("LOGIN: Authentication successful!")
-                val authResponse = result.getOrNull()
-                if (authResponse != null) {
-                    println("LOGIN: Got auth response, saving state")
-                    // Update ViewModel state
-                    _isAuthenticated = true
-                    _authToken = authResponse.authToken
-
-                    // Save to repository state management and trigger login event
-                    authRepository.saveAuthenticationState(
-                        authResponse.authToken,
-                        authResponse.userInfo
-                    )
-                    authRepository.triggerLoginEvent(authResponse.userInfo)
-                    println("LOGIN: State saved and login event triggered")
-                } else {
-                    println("LOGIN: Auth response was null")
-                }
-            } else {
-                println("LOGIN: Authentication failed")
-                val error = result.exceptionOrNull()
-                println("LOGIN: Error details: ${error?.message}")
-                // Handle error state
-                _isAuthenticated = false
-                _authToken = null
-                _errorMessage = result.exceptionOrNull()?.message ?: "Authentication failed"
-                println("LOGIN: Error message set to: $_errorMessage")
-            }
-        }.start()
-    }
-
-    fun logoutWithStateManagement() {
-        // Get current user info before clearing
-        val currentAuthState = authRepository.getPersistedAuthenticationState()
-
-        // Clear ViewModel state
-        _isAuthenticated = false
-        _authToken = null
-        _errorMessage = null
-
-        // Clear repository state and trigger logout event
-        if (currentAuthState != null) {
-            authRepository.triggerLogoutEvent(currentAuthState.userInfo)
-        }
-        authRepository.clearAuthenticationState()
-    }
-}
-
-// PLY-44: Authenticated API Client Data Models
-data class AuthenticatedRequestResult(
-    val httpResponse: String,
-    val authTokenUsed: String
-) {
-    fun containsAuthToken(token: String): Boolean {
-        return authTokenUsed == token
-    }
-}
-
-data class ServerRoutingResult(
-    val attemptedServers: List<String>,
-    val successfulServer: String?
-)
-
-// PLY-44: Authenticated API Client
-class AuthenticatedApiClient(private val authRepository: AuthRepository) {
-    private val httpClient = OkHttpClient()
-    private val gson = Gson()
-
-    fun makeAuthenticatedRequest(endpoint: String): Result<AuthenticatedRequestResult> {
-        return try {
-            // Get current authentication state
-            val authState = authRepository.getPersistedAuthenticationState()
-                ?: return Result.failure(AuthenticationException("No authentication state found"))
-
-            // Get base URL from repository
-            val baseUrl = authRepository.getLastSuccessfulServer() ?: "https://eapi.pcloud.com"
-
-            // Make authenticated request with token injection
-            val requestBody = FormBody.Builder()
-                .add("auth", authState.authToken) // Inject auth token
-                .build()
-
-            val request = Request.Builder()
-                .url("$baseUrl$endpoint")
-                .post(requestBody)
-                .build()
-
-            val response = httpClient.newCall(request).execute()
-            val responseBody = response.body?.string() ?: ""
-
-            // Create result with auth token confirmation
-            val result = AuthenticatedRequestResult(
-                httpResponse = responseBody,
-                authTokenUsed = authState.authToken
-            )
-
-            Result.success(result)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun makeRequestWithAutoRouting(endpoint: String): Result<ServerRoutingResult> {
-        val servers = listOf("https://eapi.pcloud.com", "https://api.pcloud.com")
-        val attemptedServers = mutableListOf<String>()
-        var successfulServer: String? = null
-
-        for (serverUrl in servers) {
-            attemptedServers.add(serverUrl.removePrefix("https://"))
-
-            try {
-                // Get auth state
-                val authState = authRepository.getPersistedAuthenticationState()
-                    ?: return Result.failure(
-                        AuthenticationException("No authentication state found")
-                    )
-
-                // Make request to this server
-                val requestBody = FormBody.Builder()
-                    .add("auth", authState.authToken)
-                    .build()
-
-                val request = Request.Builder()
-                    .url("$serverUrl$endpoint")
-                    .post(requestBody)
-                    .build()
-
-                val response = httpClient.newCall(request).execute()
-
-                if (response.isSuccessful) {
-                    successfulServer = serverUrl.removePrefix("https://")
-                    break
-                }
-            } catch (e: Exception) {
-                // Continue to next server
-                continue
-            }
-        }
-
-        val result = ServerRoutingResult(
-            attemptedServers = attemptedServers,
-            successfulServer = successfulServer
-        )
-
-        return Result.success(result)
-    }
-
-    fun handlePCloudErrorResponse(jsonResponse: String): Result<Nothing> {
-        return try {
-            val pCloudResponse = gson.fromJson(jsonResponse, PCloudResponse::class.java)
-
-            when (pCloudResponse.result) {
-                in 2000..2999 -> {
-                    // 2000 series - authentication errors
-                    val errorMessage = pCloudResponse.error ?: "Authentication error (${pCloudResponse.result})"
-                    Result.failure(AuthenticationException(errorMessage))
-                }
-                in 4000..4999 -> {
-                    // 4000 series - access/permission errors
-                    val errorMessage = pCloudResponse.error ?: "Access error (${pCloudResponse.result})"
-                    Result.failure(AccessException(errorMessage))
-                }
-                else -> {
-                    // Other errors
-                    val errorMessage = pCloudResponse.error ?: "Unknown error (${pCloudResponse.result})"
-                    Result.failure(Exception(errorMessage))
-                }
-            }
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-}
-
-// PLY-82: Authentication Navigation Integration
-object AuthRoutes {
-    const val Login = "Login"
-    const val Home = "Home"
-}
-
-// Navigation guard for authentication
-class AuthGuard(private val authRepository: AuthRepository) {
-    fun shouldRedirectToLogin(): Boolean {
-        return !authRepository.isAuthenticated()
-    }
-
-    fun getStartDestination(): String {
-        return if (shouldRedirectToLogin()) {
-            AuthRoutes.Login
-        } else {
-            AuthRoutes.Home
-        }
-    }
-}
-
-// Authentication-aware navigator interface
-interface AuthNavigator {
-    fun navigateToLogin()
-    fun navigateToHome()
-    fun navigateToRoute(route: String)
-}
-
-// Implementation for testing
-class TestAuthNavigator : AuthNavigator {
-    var lastNavigatedRoute: String? = null
-    var loginNavigationCount: Int = 0
-    var homeNavigationCount: Int = 0
-
-    override fun navigateToLogin() {
-        lastNavigatedRoute = AuthRoutes.Login
-        loginNavigationCount++
-    }
-
-    override fun navigateToHome() {
-        lastNavigatedRoute = AuthRoutes.Home
-        homeNavigationCount++
-    }
-
-    override fun navigateToRoute(route: String) {
-        lastNavigatedRoute = route
-    }
-}
-
-// PLY-83: Functional Login Screen with Form UI
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun LoginScreenWithNavigation(
-    authViewModel: AuthViewModel,
-    onLoginSuccess: () -> Unit
-) {
-    // PLY-83: Functional login form state - empty for production
-    var username by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-
-    // Navigation effect: when authentication succeeds, navigate to home
-    LaunchedEffect(authViewModel.isAuthenticated) {
-        if (authViewModel.isAuthenticated) {
-            onLoginSuccess()
-        }
-    }
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = "pCloud Music Player",
-            style = MaterialTheme.typography.headlineMedium,
-            color = MaterialTheme.colorScheme.primary
-        )
-
-        Spacer(modifier = Modifier.height(32.dp))
-
-        // Username/Email field
-        OutlinedTextField(
-            value = username,
-            onValueChange = { username = it },
-            label = { Text("Email or Username") },
-            modifier = Modifier.fillMaxWidth(),
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email),
-            singleLine = true
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        // Password field
-        OutlinedTextField(
-            value = password,
-            onValueChange = { password = it },
-            label = { Text("Password") },
-            modifier = Modifier.fillMaxWidth(),
-            visualTransformation = PasswordVisualTransformation(),
-            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
-            singleLine = true
-        )
-
-        Spacer(modifier = Modifier.height(24.dp))
-
-        // Login button - proper email validation
-        val emailRegex = "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$".toRegex()
-        val isValidEmail = username.matches(emailRegex)
-        val isValidUsername = username.matches("^[A-Za-z0-9_.-]+$".toRegex()) // alphanumeric, underscore, dot, dash
-        val isValidInput = isValidEmail || isValidUsername
-        val isFormValid = username.isNotBlank() && password.isNotBlank() && isValidInput
-
-        Button(
-            onClick = {
-                println("LOGIN: Button clicked, calling authViewModel.login()")
-                // Auto-detect server (tries both EU and US automatically)
-                authViewModel.login(username, password, "AUTO")
-            },
-            modifier = Modifier.fillMaxWidth(),
-            enabled = isFormValid && !authViewModel.isLoading
-        ) {
-            Text(if (authViewModel.isLoading) "Logging in..." else "Login")
-        }
-
-        // Validation message
-        if (username.isNotBlank() && !isValidInput) {
-            Spacer(modifier = Modifier.height(8.dp))
-            Text(
-                text = "Please enter a valid email address or username (letters, numbers, dots, dashes, underscores only)",
-                color = MaterialTheme.colorScheme.error,
-                style = MaterialTheme.typography.bodySmall
-            )
-        }
-
-        // Error message display
-        if (authViewModel.errorMessage != null) {
-            Spacer(modifier = Modifier.height(16.dp))
-            Text(
-                text = authViewModel.errorMessage!!,
-                color = MaterialTheme.colorScheme.error
-            )
-        }
-    }
-}
-
-// PLY-43: Secure Token Storage with Time-based Expiration
-class SecureTokenStorage {
-
-    fun storeToken(alias: String, token: String): Result<Unit> {
-        return try {
-            // For unit tests, use a secure in-memory storage simulation
-            // In real Android app, this would use Android Keystore
-            val secureStorage = getSecureStorage()
-            val expirationStorage = getExpirationStorage()
-
-            secureStorage[alias] = encryptToken(token)
-            // No expiration for regular storeToken method
-            expirationStorage.remove(alias)
-
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun storeTokenWithExpiration(alias: String, token: String, expirationDurationMs: Long): Result<Unit> {
-        return try {
-            val secureStorage = getSecureStorage()
-            val expirationStorage = getExpirationStorage()
-            val activityStorage = getActivityTimeoutStorage()
-
-            secureStorage[alias] = encryptToken(token)
-            expirationStorage[alias] = System.currentTimeMillis() + expirationDurationMs
-            // Clear activity timeout for fixed expiration tokens
-            activityStorage.remove(alias)
-
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun storeTokenWithActivityTimeout(alias: String, token: String, inactivityTimeoutMs: Long): Result<Unit> {
-        return try {
-            val secureStorage = getSecureStorage()
-            val expirationStorage = getExpirationStorage()
-            val activityStorage = getActivityTimeoutStorage()
-
-            secureStorage[alias] = encryptToken(token)
-            // Store inactivity timeout duration and last access time
-            activityStorage[alias] = ActivityInfo(inactivityTimeoutMs, System.currentTimeMillis())
-            // Clear fixed expiration for activity-based tokens
-            expirationStorage.remove(alias)
-
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun storeTokenWithAutoRefresh(
-        alias: String,
-        token: String,
-        expirationDurationMs: Long,
-        refreshThresholdMs: Long,
-        refreshFunction: (String) -> Result<String>
-    ): Result<Unit> {
-        return try {
-            val secureStorage = getSecureStorage()
-            val expirationStorage = getExpirationStorage()
-            val activityStorage = getActivityTimeoutStorage()
-            val refreshStorage = getRefreshStorage()
-
-            secureStorage[alias] = encryptToken(token)
-            expirationStorage[alias] = System.currentTimeMillis() + expirationDurationMs
-            // Store refresh configuration
-            refreshStorage[alias] = RefreshInfo(
-                refreshThresholdMs,
-                expirationDurationMs,
-                refreshFunction
-            )
-            // Clear activity timeout for auto-refresh tokens
-            activityStorage.remove(alias)
-
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun storeTokenWithValidation(
-        alias: String,
-        token: String,
-        validationFunction: (String) -> Boolean
-    ): Result<Unit> {
-        return try {
-            // Validate token before storing
-            if (!validationFunction(token)) {
-                return Result.failure(
-                    TokenValidationException("Token validation failed for alias: $alias")
-                )
-            }
-
-            val secureStorage = getSecureStorage()
-            val validationStorage = getValidationStorage()
-
-            secureStorage[alias] = encryptToken(token)
-            // Store validation function for later use
-            validationStorage[alias] = validationFunction
-
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun simulateTokenCorruption(alias: String) {
-        // Simulate token corruption by completely replacing the encrypted token with garbage
-        val secureStorage = getSecureStorage()
-        if (secureStorage.containsKey(alias)) {
-            // Replace with corrupted data that will decrypt to something that fails validation
-            secureStorage[alias] = "CORRUPTED_DATA_THAT_FAILS_VALIDATION_ENCRYPTED"
-        }
-    }
-
-    fun removeToken(alias: String): Result<Unit> {
-        return try {
-            val secureStorage = getSecureStorage()
-            val expirationStorage = getExpirationStorage()
-            val activityStorage = getActivityTimeoutStorage()
-            val refreshStorage = getRefreshStorage()
-            val validationStorage = getValidationStorage()
-
-            // Remove from all storage types
-            secureStorage.remove(alias)
-            expirationStorage.remove(alias)
-            activityStorage.remove(alias)
-            refreshStorage.remove(alias)
-            validationStorage.remove(alias)
-
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun retrieveToken(alias: String): Result<String> {
-        return try {
-            val secureStorage = getSecureStorage()
-            val expirationStorage = getExpirationStorage()
-            val activityStorage = getActivityTimeoutStorage()
-            val refreshStorage = getRefreshStorage()
-            val validationStorage = getValidationStorage()
-
-            // Check if token exists
-            val encryptedToken = secureStorage[alias] ?: return Result.failure(
-                IllegalArgumentException("Token not found for alias: $alias")
-            )
-
-            val currentTime = System.currentTimeMillis()
-
-            // Check if token has fixed expiration
-            val fixedExpirationTime = expirationStorage[alias]
-            if (fixedExpirationTime != null) {
-                // Check if token has auto-refresh capability
-                val refreshInfo = refreshStorage[alias]
-                if (refreshInfo != null) {
-                    // Token has auto-refresh - check if we need to refresh
-                    val timeUntilExpiration = fixedExpirationTime - currentTime
-                    if (timeUntilExpiration <= refreshInfo.refreshThresholdMs) {
-                        // Need to refresh token
-                        val currentToken = decryptToken(encryptedToken)
-                        val refreshResult = refreshInfo.refreshFunction(currentToken)
-
-                        if (refreshResult.isSuccess) {
-                            val newToken = refreshResult.getOrNull()!!
-                            // Store refreshed token with new expiration (same duration as original)
-                            secureStorage[alias] = encryptToken(newToken)
-                            expirationStorage[alias] = currentTime + refreshInfo.originalDurationMs
-
-                            return Result.success(newToken)
-                        } else {
-                            // Refresh failed - return failure
-                            return Result.failure(
-                                Exception(
-                                    "Token refresh failed: ${refreshResult.exceptionOrNull()?.message}"
-                                )
-                            )
-                        }
-                    }
-                } else {
-                    // Token has fixed expiration without auto-refresh - check if it's expired
-                    if (currentTime > fixedExpirationTime) {
-                        // Token is expired - remove it and throw exception
-                        secureStorage.remove(alias)
-                        expirationStorage.remove(alias)
-                        return Result.failure(
-                            TokenExpiredException("Token expired for alias: $alias")
-                        )
-                    }
-                }
-            }
-
-            // Check if token has activity-based expiration
-            val activityInfo = activityStorage[alias]
-            if (activityInfo != null) {
-                // Token has activity timeout - check if it's expired due to inactivity
-                val timeSinceLastAccess = currentTime - activityInfo.lastAccessTime
-                if (timeSinceLastAccess > activityInfo.timeoutMs) {
-                    // Token is expired due to inactivity - remove it and throw exception
-                    secureStorage.remove(alias)
-                    activityStorage.remove(alias)
-                    return Result.failure(
-                        TokenExpiredException("Token expired due to inactivity for alias: $alias")
-                    )
-                }
-
-                // Token is still valid - extend the activity timeout by updating last access time
-                activityStorage[alias] = activityInfo.copy(lastAccessTime = currentTime)
-            }
-
-            val decryptedToken = try {
-                decryptToken(encryptedToken)
-            } catch (e: Exception) {
-                // Decryption failed - token is corrupted
-                secureStorage.remove(alias)
-                return Result.failure(
-                    TokenValidationException(
-                        "Token decryption failed for alias: $alias - token may be corrupted"
-                    )
-                )
-            }
-
-            // Check if token has validation requirements
-            val validationFunction = validationStorage[alias]
-            if (validationFunction != null) {
-                // Validate token before returning
-                if (!validationFunction(decryptedToken)) {
-                    // Token validation failed - remove it and throw exception
-                    secureStorage.remove(alias)
-                    validationStorage.remove(alias)
-                    return Result.failure(
-                        TokenValidationException("Token validation failed for alias: $alias")
-                    )
-                }
-            }
-
-            Result.success(decryptedToken)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    // Simulate secure storage (in production, this would be Android Keystore)
-    private fun getSecureStorage(): MutableMap<String, String> {
-        return tokenStorage
-    }
-
-    // Simulate expiration storage (in production, this would be part of Android Keystore metadata)
-    private fun getExpirationStorage(): MutableMap<String, Long> {
-        return expirationStorage
-    }
-
-    // Simulate activity timeout storage (in production, this would be part of Android Keystore metadata)
-    private fun getActivityTimeoutStorage(): MutableMap<String, ActivityInfo> {
-        return activityTimeoutStorage
-    }
-
-    // Simulate refresh storage (in production, this would be part of Android Keystore metadata)
-    private fun getRefreshStorage(): MutableMap<String, RefreshInfo> {
-        return refreshStorage
-    }
-
-    // Simulate validation storage (in production, this would be part of Android Keystore metadata)
-    private fun getValidationStorage(): MutableMap<String, (String) -> Boolean> {
-        return validationStorage
-    }
-
-    // Data class to store activity-based expiration information
-    private data class ActivityInfo(
-        val timeoutMs: Long, // How long token stays valid without activity
-        val lastAccessTime: Long // When token was last accessed
-    )
-
-    // Data class to store auto-refresh information
-    private data class RefreshInfo(
-        val refreshThresholdMs: Long, // When to trigger refresh before expiration
-        val originalDurationMs: Long, // Original token duration for refresh calculation
-        val refreshFunction: (String) -> Result<String> // Function to call for token refresh
-    )
-
-    // Simulate encryption (in production, this would use Android Keystore encryption)
-    private fun encryptToken(token: String): String {
-        // Simple obfuscation for unit test (production would use real encryption)
-        return token.reversed() + "_ENCRYPTED"
-    }
-
-    // Simulate decryption (in production, this would use Android Keystore decryption)
-    private fun decryptToken(encryptedToken: String): String {
-        // Reverse the simple obfuscation for unit test
-        return encryptedToken.removeSuffix("_ENCRYPTED").reversed()
-    }
-
-    companion object {
-        // Simulate secure storage (in production, this would be Android Keystore)
-        private val tokenStorage = mutableMapOf<String, String>()
-
-        // Simulate expiration times storage
-        private val expirationStorage = mutableMapOf<String, Long>()
-
-        // Simulate activity timeout storage
-        private val activityTimeoutStorage = mutableMapOf<String, ActivityInfo>()
-
-        // Simulate refresh configuration storage
-        private val refreshStorage = mutableMapOf<String, RefreshInfo>()
-
-        // Simulate validation function storage
-        private val validationStorage = mutableMapOf<String, (String) -> Boolean>()
-    }
-}
+// Direct class exports for backward compatibility - preserved original patterns
+val AuthRepository = com.foxy.player.authentication.network.AuthRepository::class
+val AuthenticatedApiClient = com.foxy.player.authentication.network.AuthenticatedApiClient::class
+val AuthScreen = com.foxy.player.authentication.ui.AuthScreen::class
+val LoginScreen = com.foxy.player.authentication.ui.LoginScreen::class
+val AuthViewModel = com.foxy.player.authentication.ui.AuthViewModel::class
+val SecureTokenStorage = com.foxy.player.authentication.utils.SecureTokenStorage::class
