+++ app/src/main/java/com/foxy/player/music/MusicHubViewModel.kt
@@ -2,8 +2,8 @@ package com.foxy.player.music
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.foxy.player.authentication.AuthRepository
-import com.foxy.player.authentication.AuthenticatedApiClient
+import com.foxy.player.authentication.network.AuthRepository
+import com.foxy.player.authentication.network.AuthenticatedApiClient
 import com.foxy.player.music.network.MusicDiscoveryService
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
