+++ app/src/main/java/com/foxy/player/music/MusicLibrary.kt
@@ -17,7 +17,7 @@ import androidx.work.PeriodicWorkRequestBuilder
 import androidx.work.WorkManager
 import androidx.work.WorkerParameters
 import androidx.work.workDataOf
-import com.foxy.player.authentication.AuthenticatedApiClient
+import com.foxy.player.authentication.network.AuthenticatedApiClient
 import com.foxy.player.music.entities.AudioFile
 import com.foxy.player.music.sync.*
 import com.foxy.player.music.ui.*
