+++ app/src/main/java/com/foxy/player/music/MusicSearch.kt
@@ -1,6 +1,6 @@
 package com.foxy.player.music
 
-import com.foxy.player.authentication.AuthenticatedApiClient
+import com.foxy.player.authentication.network.AuthenticatedApiClient
 import com.foxy.player.music.ui.*
 
 // ===== MUSIC SEARCH AND BROWSE SERVICE =====
