+++ app/src/test/java/com/foxy/player/authentication/AuthenticatedApiClientTest.kt
@@ -1,5 +1,8 @@
 package com.foxy.player.authentication
 
+import com.foxy.player.authentication.models.*
+import com.foxy.player.authentication.network.AuthRepository
+import com.foxy.player.authentication.network.AuthenticatedApiClient
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
