+++ music-player-core/PLY-135-refactor-authentication-kloc-file.md
@@ -1,0 +1,56 @@
+# Refactor Authentication KLOC File
+
+**Objective**
+Break down auth.kt (1340 lines) into business/, network/, utils/, models/ modules - separate authentication concerns while maintaining single file per domain
+
+**Implementation Notes**
+
+* Follow test-driven development approach
+* Keep all domain code in single file for AI context optimization
+* Use existing codebase patterns and conventions
+
+**Definition of Done**
+
+* Tests written and passing
+* Core functionality implemented
+* Integration validated with existing systems
+* Code review completed
+
+---
+## Android Development Context (Auto-Generated)
+
+**Project**: pCloud Music Player (Android)  
+**Technology Stack**: 
+- Language: Kotlin
+- UI Framework: Jetpack Compose
+- Architecture: MVVM with Repository pattern
+- Audio Engine: Media3/ExoPlayer
+- Networking: Retrofit + OkHttp
+- DI: Hilt/Dagger
+- Build System: Gradle
+
+**Development Domain**: music-player-core  
+**Category**: general  
+
+**Key Android Components Likely Needed**:
+- Activities/Fragments for UI
+- Services for background playback
+- Repository for data access (pCloud API)
+- ViewModels for UI state management
+- Compose UI components
+- MediaSession for media controls
+
+**Testing Strategy**:
+- Unit tests with JUnit 5 + Mockito
+- UI tests with Compose Test
+- Integration tests for API calls
+- Instrumented tests for Android components
+
+**Quality Tools**:
+- ktlint for Kotlin style
+- detekt for code quality
+- Android Lint for security/performance
+- Gradle for build validation
+
+---
+*Generated by Spectrum Development Framework for Android*
