+++ music-player-core/tickets/PLY-134-refactor-musicdiscoverykt-kloc-file.md
@@ -54,3 +54,17 @@ Split MusicDiscovery.kt (1132 lines) into business/, network/, utils/ modules - 
 
 ---
 *Generated by Spectrum Development Framework for Android*
+
+---
+## Implementation Completed
+- **Ticket**: PLY-134
+- **PR**: https://github.com/appmancer/FoxyPlayer/pull/43
+- **Domain**: music-player-core
+- **TDD Cycles**: 0 completed
+- **Tests**: 0 passing
+- **Files Changed**: 20
+- **Merged**: 2025-09-02T13:03:14+01:00
+- **Branch**: feature/PLY-134-refactor-musicdiscoverykt-kloc-file (deleted)
+
+This ticket has been completed and deployed through the Centro development workflow.
+The implementation has been merged to dev branch and deployed to staging environment.
